try add DependencyNodeAnalyzer
