--- packages/coding-standard/src/NodeAnalyzer/InsideAbstractClassAndPassedAsDependencyViaConstructorAnalizer.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Symplify\CodingStandard\NodeAnalyzer;
-
-final class InsideAbstractClassAndPassedAsDependencyViaConstructorAnalizer
-{
-
-}
